+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.31.2] - 2023-03-07
+
+### Fixes
+
+-   Keep redirectToPath's query params while redirecting to it from auth
+
 ### Changes
 
 -   Update example apps after header based auth